+++ app/src/main/java/com/eps/todoturtle/MainActivity.kt
@@ -3,9 +3,9 @@ package com.eps.todoturtle
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.eps.todoturtle.permissions.PermissionRequester
-import com.eps.todoturtle.permissions.RequestPermissionContext
-import com.eps.todoturtle.permissions.providers.CameraPermissionProvider
+import com.eps.todoturtle.permissions.logic.PermissionRequester
+import com.eps.todoturtle.permissions.logic.RequestPermissionContext
+import com.eps.todoturtle.permissions.logic.providers.CameraPermissionProvider
 import com.eps.todoturtle.ui.MainActivityUI
 import com.eps.todoturtle.ui.theme.ToDoTurtleTheme
