+++ app/src/main/java/com/eps/todoturtle/permissions/logic/providers/CameraPermissionProvider.kt
@@ -1,7 +1,7 @@
-package com.eps.todoturtle.permissions.providers
+package com.eps.todoturtle.permissions.logic.providers
 
 import android.Manifest
-import com.eps.todoturtle.permissions.providers.text.CameraTextProvider
+import com.eps.todoturtle.permissions.logic.providers.text.CameraTextProvider
 
 class CameraPermissionProvider : PermissionProvider {
     override val textProvider = CameraTextProvider()
