+++ app/src/main/java/com/eps/todoturtle/permissions/logic/providers/PermissionProvider.kt
@@ -1,0 +1,8 @@
+package com.eps.todoturtle.permissions.logic.providers
+
+import com.eps.todoturtle.permissions.logic.providers.text.PermissionTextProvider
+
+interface PermissionProvider {
+    val textProvider: PermissionTextProvider
+    val permission: String
+}
