+++ app/src/main/java/com/eps/todoturtle/permissions/logic/providers/text/CameraTextProvider.kt
@@ -1,4 +1,4 @@
-package com.eps.todoturtle.permissions.providers.text
+package com.eps.todoturtle.permissions.logic.providers.text
 
 class CameraTextProvider : PermissionTextProvider {
     override val isPermanentlyDeclined: String
