--- app/src/main/java/com/eps/todoturtle/permissions/providers/PermissionProvider.kt
@@ -1,8 +1,0 @@
-package com.eps.todoturtle.permissions.providers
-
-import com.eps.todoturtle.permissions.providers.text.PermissionTextProvider
-
-interface PermissionProvider {
-    val textProvider: PermissionTextProvider
-    val permission: String
-}
