+++ app/src/main/java/com/eps/todoturtle/profile/logic/HostageType.kt
@@ -1,4 +1,4 @@
-package com.eps.todoturtle.profile
+package com.eps.todoturtle.profile.logic
 
 import android.content.Context
 import android.os.Parcelable
