+++ app/src/main/java/com/eps/todoturtle/profile/logic/ProfileDetails.kt
@@ -1,4 +1,4 @@
-package com.eps.todoturtle.profile
+package com.eps.todoturtle.profile.logic
 
 import android.graphics.Bitmap
 import androidx.compose.runtime.MutableState
