+++ app/src/main/java/com/eps/todoturtle/shared/logic/exceptions/DrawableNotFoundException.kt
@@ -1,3 +1,3 @@
-package com.eps.todoturtle.exceptions
+package com.eps.todoturtle.shared.logic.exceptions
 
 class DrawableNotFoundException(message: String) : Exception(message)
